+++ NEWS.md
@@ -36,6 +36,7 @@ SQL Server (#75)
 - `db_analyze()` unsupported and simply returns `TRUE`.
 - `src_sqlserver` has a nicer print
 - `intersect()` and `setdiff()` methods are deprecated and default `tbl_sql` methods from `dplyr` are called instead.
+- Added basic testing of dplyr backend (#81)
 
 ## Other changes
 
@@ -47,6 +48,7 @@ SQL Server (#75)
 - Introduced `pattern` argument to `dbListTables()` which allows you to list all tables matching a pattern.
 - `dbExistsTable()` now passed table name to `dbListTables()` as a pattern to be matched which should improve its performance.
 - `dbGetInfo()` for `SQLServerResult` has been deprecated and calls the DBI default method which calls `dbHasCompleted()`, `dbGetRowCount()` etc. The latter methods have been implemented for `SQLServerResult` and are exported.
+- `as.numeric()` and `as.character()` calls now cast scalar input values to SQL types `FLOAT` and `NVARCHAR(4000)` respectively rather than `NUMERIC` and `TEXT` respectively (default in dplyr). 
 - Now rely on DBI supplied `show()` methods
 - Added Travis-CI support